Add temp token
